delete the update package,add log4j2.
