+++ demo.xml
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="utf-8"?><patent name="mypatent"><title>woshi</title><claim claims="any">woshiaclaim</claim></patent>
