+++ src/com/icdd/test/TestLucence.java
@@ -3,13 +3,13 @@ package com.icdd.test;
 import org.junit.Test;
 
 
-import com.icdd.lucence.LucenceIndex;
+import com.icdd.lucence.IndexFiles;
 import com.icdd.lucence.SearchFiles;
 
 public class TestLucence {
 	@Test
 	public void indexFiles(){
-		LucenceIndex indexTest = new LucenceIndex();
+		IndexFiles indexTest = new IndexFiles();
 		indexTest.index();
 	}
 	@Test
