--- src/uptest/Test.java
@@ -1,7 +1,0 @@
-package uptest;
-
-public class Test {
-  public static void main(String[] args){
-       System.out.print("hello world!");
-  }
-}
